update trainer docs
